requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.0
SQLAlchemy==2.0.21
pydantic==2.4.0
openai
python-dotenv